+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-client/src/test/java/org/allGraphQLCases/subscription/ExecSubscriptionIT.java
@@ -1,9 +1,9 @@
 package org.allGraphQLCases.subscription;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
@@ -26,6 +26,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.graphql.client.GraphQlTransportException;
 
 import com.graphql_java_generator.client.SubscriptionClient;
 import com.graphql_java_generator.exception.GraphQLRequestExecutionException;
@@ -236,7 +237,7 @@ public class ExecSubscriptionIT {
 	 */
 	@Test
 	@Execution(ExecutionMode.CONCURRENT)
-	public void test_subscribeToADate_serverComplete()
+	public void test_subscriptionTest_serverComplete()
 			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException, InterruptedException {
 		logger.info("------------------------------------------------------------------------------------------------");
 		logger.info("Starting test_subscribeToADate_serverComplete");
@@ -270,7 +271,7 @@ public class ExecSubscriptionIT {
 	 */
 	@Test
 	@Execution(ExecutionMode.CONCURRENT)
-	public void test_subscribeToADate_clientComplete()
+	public void test_subscriptionTest_clientComplete()
 			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException, InterruptedException {
 		logger.info("------------------------------------------------------------------------------------------------");
 		logger.info("Starting test_subscribeToADate_clientComplete");
@@ -300,12 +301,17 @@ public class ExecSubscriptionIT {
 
 	@Test
 	@Execution(ExecutionMode.CONCURRENT)
-	void test_connectionError() throws GraphQLRequestExecutionException, GraphQLRequestPreparationException {
+	void test_connectionError()
+			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException, InterruptedException {
 		Date date = new Calendar.Builder().setDate(2018, 02, 01).build().getTime();
 		SubscriptionCallbackToADate callback = new SubscriptionCallbackToADate("test_connectionError");
-		GraphQLRequestExecutionException e = assertThrows(GraphQLRequestExecutionException.class,
-				() -> subscriptionExecutor2.issue53("", callback, date));
-		assertTrue(e.getMessage().contains("Connection refused"), "The received error message is: " + e.getMessage());
+		SubscriptionClient sub = subscriptionExecutor2.issue53("", callback, date);
+		callback.latchForMessageReception.await(20, TimeUnit.SECONDS);
+		assertInstanceOf(GraphQlTransportException.class, callback.lastReceivedError);
+		assertTrue(callback.lastReceivedError.getMessage().contains("Connection refused"),
+				"The received error message is: " + callback.lastReceivedError.getMessage());
+
+		sub.unsubscribe();
 	}
 
 	/**
@@ -318,7 +324,7 @@ public class ExecSubscriptionIT {
 	 */
 	@Test
 	@Execution(ExecutionMode.CONCURRENT)
-	public void test_subscribeToADate_subscriptionError()
+	public void test_subscriptionTest_subscriptionError()
 			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException, InterruptedException {
 		logger.info("------------------------------------------------------------------------------------------------");
 		logger.info("Starting test_subscribeToADate_subscriptionError");
@@ -335,7 +341,7 @@ public class ExecSubscriptionIT {
 		assertNotNull(callback.lastExceptionReceived, "we should have received an exception");
 		assertTrue(
 				callback.lastExceptionReceived.getMessage()
-						.endsWith("Oups, the subscriber asked for an error during the subscription"),
+						.contains("Oups, the subscriber asked for an error during the subscription"),
 				"The received error message is: " + callback.lastExceptionReceived.getMessage());
 
 		// Let's unsubscribe from this subscription
@@ -352,7 +358,7 @@ public class ExecSubscriptionIT {
 	 */
 	@Test
 	@Execution(ExecutionMode.CONCURRENT)
-	public void test_subscribeToADate_nextError()
+	public void test_subscriptionTest_nextError()
 			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException, InterruptedException {
 		logger.info("------------------------------------------------------------------------------------------------");
 		logger.info("Starting test_subscribeToADate_nextError");
@@ -369,7 +375,7 @@ public class ExecSubscriptionIT {
 		assertNotNull(callback.lastExceptionReceived, "we must have received an exception");
 		assertTrue(
 				callback.lastExceptionReceived.getMessage()
-						.endsWith("Oups, the subscriber asked for an error for each next message"),
+						.contains("Oups, the subscriber asked for an error for each next message"),
 				"The received error message is: " + callback.lastExceptionReceived.getMessage());
 
 		// Let's unsubscribe from this subscription
@@ -386,7 +392,7 @@ public class ExecSubscriptionIT {
 	 */
 	@Test
 	@Execution(ExecutionMode.CONCURRENT)
-	public void test_subscribeToADate_webSocketCloseError()
+	public void test_subscriptionTest_webSocketCloseError()
 			throws GraphQLRequestExecutionException, GraphQLRequestPreparationException, InterruptedException {
 		logger.info("------------------------------------------------------------------------------------------------");
 		logger.info("Starting test_subscribeToADate_webSocketCloseError");
@@ -402,9 +408,9 @@ public class ExecSubscriptionIT {
 
 		// Let's test this exception
 		assertNotNull(callback.lastExceptionReceived, "we must have received an exception");
-		assertTrue(callback.lastExceptionReceived.getMessage().endsWith(
-				"Oups, the subscriber asked that the web socket get disconnected before the first notification"),
-				"The received error message is: " + callback.lastExceptionReceived.getMessage());
+		assertTrue(callback.lastExceptionReceived.getMessage().contains(
+				"message=Oups, the subscriber asked that the web socket get disconnected before the first notification"));
+		assertTrue(callback.lastExceptionReceived.getMessage().contains("classification=BAD_REQUEST"));
 
 		// Let's unsubscribe from this subscription
 		sub.unsubscribe();
@@ -449,7 +455,7 @@ public class ExecSubscriptionIT {
 		// Let's test this exception
 		assertNull(callback.lastExceptionReceived, "we must have received no exception");
 		assertNotNull(callback.lastReceivedMessage, "we must have received a message");
-		assertEquals(3, callback.lastReceivedMessage.size(),
+		assertEquals(4, callback.lastReceivedMessage.size(),
 				"each received notifiation should contain a list of 3 items");
 		assertEquals(CEP_EnumWithReservedJavaKeywordAsValues_CES._int, callback.lastReceivedMessage.get(0),
 				"First item should be the 'int' value of the enum");
@@ -457,6 +463,7 @@ public class ExecSubscriptionIT {
 				"Second item should be the 'interface' value of the enum");
 		assertEquals(CEP_EnumWithReservedJavaKeywordAsValues_CES._long, callback.lastReceivedMessage.get(2),
 				"Third item should be the 'long' value of the enum");
+		assertNull(callback.lastReceivedMessage.get(3), "Fourth item is null");
 
 		// Let's unsubscribe from this subscription
 		sub.unsubscribe();