+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-server/src/main/java/org/allGraphQLCases/server/config/GraphQlException.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package org.allGraphQLCases.server.extensions;
+package org.allGraphQLCases.server.config;
 
 import java.util.List;