+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-server/src/main/java/org/allGraphQLCases/server/config/MyInstrumentation.java
@@ -1,4 +1,4 @@
-package org.allGraphQLCases.server.extensions;
+package org.allGraphQLCases.server.config;
 
 import java.util.Collections;
 import java.util.LinkedHashMap;