+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-server/src/main/java/org/allGraphQLCases/server/config/MySubscriptionExceptionResolverAdapter.java
@@ -1,0 +1,25 @@
+package org.allGraphQLCases.server.config;
+
+import org.springframework.graphql.execution.SubscriptionExceptionResolverAdapter;
+import org.springframework.stereotype.Component;
+
+import graphql.GraphQLError;
+
+/**
+ * Implements the spring-graphql exception management, for Subscriptions
+ * 
+ * @author etienne-sf
+ */
+@Component
+public class MySubscriptionExceptionResolverAdapter extends SubscriptionExceptionResolverAdapter {
+
+	@Override
+	protected GraphQLError resolveToSingleError(Throwable exception) {
+		if (exception instanceof GraphQlException) {
+			return (GraphQLError) exception;
+		} else {
+			return null;
+		}
+	}
+
+}