+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-server/src/main/java/org/allGraphQLCases/server/extensions/package-info.java
@@ -1,5 +1,0 @@
-package org.allGraphQLCases.server.extensions;
-
-/**
- * This package contains the classes that allows to fill the extensions attribute of the GraphQL response
- */